pyproject.toml
[project]
name = "upvc-club"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-dynamic-preferences>=1.17.0",
    "django>=6.0.1",
    "pillow>=12.1.0",
    "requests>=2.32.5",
    "weasyprint>=68.0",
    "django-pwa>=2.0.1",
    "python-dotenv>=1.2.1",
    "gunicorn>=25.0.3",
    "psycopg2-binary>=2.9.11",
]
